% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_binning.R
\name{binning_categorical}
\alias{binning_categorical}
\title{Bin categorical variable}
\usage{
binning_categorical(x, y, dftrain, dftest, groups = NULL)
}
\arguments{
\item{x}{Character string indicating which variable is being binned}

\item{y}{Binary response variable}

\item{dftrain}{Data frame containing the training data}

\item{dftest}{Data frame containing the test data}

\item{groups}{Pre-determined vector for binning}
}
\value{
A list of two components, the first component is the binned variable
    in the training data and the second component is the binning applied on
    the test data
}
\description{
Bin categorical variable
}
