% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_binning.R
\name{binning_numeric}
\alias{binning_numeric}
\title{Bin numeric variable}
\usage{
binning_numeric(x, y, dftrain, dftest, cuts = NULL, type = "CIT")
}
\arguments{
\item{x}{Character string indicating which variable is being binned}

\item{y}{Binary response variable}

\item{dftrain}{Data frame containing the training data}

\item{dftest}{Data frame containing the test data}

\item{cuts}{Vector of pre-determined cut points for binning}

\item{type}{Whether to bin using conditional inference trees or use quantiles}
}
\value{
A list of two components, the first component is the binned variable
    in the training data and the second component is the binning applied on
    the test data
}
\description{
Bin numeric variable
}
