% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_binning.R
\name{binning_server}
\alias{binning_server}
\title{Binning module server-side processing}
\usage{
binning_server(input, output, session, modeltraintest, modeldataspecs,
  goodbad_var)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{modeltraintest}{data frame containing the credit scoring data split
into training and testing data}

\item{modeldataspecs}{data specification table for credit scoring data}

\item{goodbad_var}{good/bad variable}
}
\value{
list with following components
\describe{
  \item{binning}{list containing binning of variables created by user}
}
}
\description{
Binning module server-side processing
}
