% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_binning.R
\name{binning_ui}
\alias{binning_ui}
\title{Interface to bin the variables for scorecard development}
\usage{
binning_ui(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Interface to bin the variables for scorecard development
}
