% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{data_generate_specs}
\alias{data_generate_specs}
\title{Return the data specification table given a data frame}
\usage{
data_generate_specs(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame comprising the variable names in the input,
    variable type (as detected by the function 'class'),
    logical to identify a particular variable as the good/bad flag,
    logical to identify a particular variable to be used for stratified
    sampling, and a
    logical to identify a particular variable to be considered for binning
}
\description{
Return the data specification table given a data frame
}
