% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{data_server}
\alias{data_server}
\title{Data upload and specification module server-side processing}
\usage{
data_server(input, output, session)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}
}
\value{
list with following components
\describe{
  \item{modeldata}{reactive data frame of data uploaded by user}
  \item{modeldataspecs}{reactive data frame of data specifications}
}
}
\description{
Data upload and specification module server-side processing
}
