% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{data_ui}
\alias{data_ui}
\title{Interface to upload model data and modify data specifications}
\usage{
data_ui(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Interface to upload model data and modify data specifications
}
