% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_download.R
\name{download_server}
\alias{download_server}
\title{Download module server-side processing}
\usage{
download_server(input, output, session, binning, scorecard, validation)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{binning}{binning object created in binning module}

\item{scorecard}{scorecard object created in model module}

\item{validation}{validation object created in validation module}
}
\description{
Download module server-side processing
}
