% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_download.R
\name{download_ui}
\alias{download_ui}
\title{Interface to download relevant datasets}
\usage{
download_ui(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Interface to download relevant datasets
}
