% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_explore.R
\name{explore_countperc_categorical}
\alias{explore_countperc_categorical}
\title{Frequency and percentage of values for a categorical variable}
\usage{
explore_countperc_categorical(x, max_values = 30)
}
\arguments{
\item{x}{A vector of character values}

\item{max_values}{Maximum number of distinct values to return; if the number
of values exceeds this then the remaining are clubbed as OTHERS}
}
\value{
A data frame of frequencies and percentages for each value
}
\description{
Frequency and percentage of values for a categorical variable
}
