% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_explore.R
\name{explore_descstat_numeric}
\alias{explore_descstat_numeric}
\title{Descriptive statistics for a numeric variable}
\usage{
explore_descstat_numeric(x, round_dec = 2)
}
\arguments{
\item{x}{A vector of numeric values}

\item{round_dec}{Number of decimal places to round the results}
}
\value{
A named vector of descriptive statistics
}
\description{
Descriptive statistics for a numeric variable
}
