% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_explore.R
\name{explore_distribution_categorical}
\alias{explore_distribution_categorical}
\title{Distribution plot for a categorical vector}
\usage{
explore_distribution_categorical(x, vname = "categorical")
}
\arguments{
\item{x}{A vector of character values}
}
\value{
A list containing the ggplot2 object showing the distribution of
    the variable as well as the data frame of frequencies and percentages
}
\description{
Distribution plot for a categorical vector
}
