% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_explore.R
\name{explore_distribution_numeric}
\alias{explore_distribution_numeric}
\title{Distribution plot for a numeric vector}
\usage{
explore_distribution_numeric(x, vname = "numeric")
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
A ggplot2 object showing the distribution of the variable
}
\description{
Distribution plot for a numeric vector
}
