% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_explore.R
\name{explore_modes}
\alias{explore_modes}
\title{Modes of a numeric or categorical variable}
\usage{
explore_modes(x, na.rm = TRUE, max_modes = 5L)
}
\arguments{
\item{x}{A vector of numeric or categorical values}

\item{na.rm}{Logical to indicate if missing values should be removed}

\item{max_modes}{Maximum number of modes to return, if there are multiple}
}
\value{
A vector containing the modes of x
}
\description{
Modes of a numeric or categorical variable
}
