% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_explore.R
\name{explore_server}
\alias{explore_server}
\title{Explore data module server-side processing}
\usage{
explore_server(input, output, session, modeldata, modeldataspecs)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{modeldata}{data frame containing the credit scoring data}

\item{modeldataspecs}{data specification table for credit scoring data}
}
\description{
Explore data module server-side processing
}
