% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_explore.R
\name{explore_ui}
\alias{explore_ui}
\title{Interface to explore the data based on data specifications}
\usage{
explore_ui(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Interface to explore the data based on data specifications
}
