% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_model.R
\name{model_add_predictions}
\alias{model_add_predictions}
\title{Add good rate, bad rate and good-bad log adds to a dataset}
\usage{
model_add_predictions(df, logistic, scale_slope, scale_intercept,
  pgood = "p_good", pbad = "p_bad", ln_gb = "log_gb_odds",
  score = "score")
}
\arguments{
\item{df}{Data set where to add predictions}

\item{logistic}{A logistic model object created using glm}

\item{scale_slope}{The slope to convert the log odds to a score}

\item{scale_intercept}{The intercept to convert the log odds to a score}

\item{pgood}{Name of the variable for good rate}

\item{pbad}{Name of the variable for bad rate}

\item{ln_gb}{Name of the variable for good-bad log odds}

\item{score}{Name of the final score variable}
}
\value{
Dataset after adding the good rate, bad rate, good-bad log odds
    and score
}
\description{
Add good rate, bad rate and good-bad log adds to a dataset
}
