% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_model.R
\name{model_scaling_parameters}
\alias{model_scaling_parameters}
\title{Return the slope and intercept of the scaling function}
\usage{
model_scaling_parameters(pdo, score, odds)
}
\arguments{
\item{pdo}{Points to double the odds}

\item{score}{Base score}

\item{odds}{Odds at base score}
}
\value{
A vector containing the scaling slope and intercept
}
\description{
Return the slope and intercept of the scaling function
}
