% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_model.R
\name{model_score_table}
\alias{model_score_table}
\title{Construct the score points for a single variable}
\usage{
model_score_table(bin, coef, intcpt, scale_slope, scale_intercept,
  n_predictors, vname)
}
\arguments{
\item{bin}{The binning object created for this variable}

\item{coef}{The logistic coefficient for this variable}

\item{intcpt}{The logistic intercept}

\item{scale_slope}{The slope to convert the log odds to a score}

\item{scale_intercept}{The intercept to convert the log odds to a score}

\item{n_predictors}{The number of predictors in the model}

\item{vname}{The name of the variable}
}
\value{
Data frame of bin values and score points
}
\description{
Construct the score points for a single variable
}
