% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_model.R
\name{model_server}
\alias{model_server}
\title{Model module server-side processing}
\usage{
model_server(input, output, session, modeltraintest, modeldataspecs,
  binning, goodbad_var)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{modeltraintest}{data frame containing the credit scoring data split
into training and testing data}

\item{modeldataspecs}{data specification table for credit scoring data}

\item{binning}{bin specifications for all variables}

\item{goodbad_var}{good/bad variable}
}
\value{
list with following components
\describe{
  \item{scorecard}{Scorecard object}
}
}
\description{
Model module server-side processing
}
