% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_model.R
\name{model_ui}
\alias{model_ui}
\title{Interface to build the scorecard}
\usage{
model_ui(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Interface to build the scorecard
}
