% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_model.R
\name{model_woe_numeric}
\alias{model_woe_numeric}
\title{Create the WoEs in the data for a numeric variable}
\usage{
model_woe_numeric(df, bin, var, newvar)
}
\arguments{
\item{df}{Data set where the WoE for the variable is to be created}

\item{bin}{The binning object created for this variable}

\item{var}{Name of the variable}

\item{newvar}{Name of the variable where the WoE is to be stored}
}
\value{
Dataset with the WoE created in a new variable
}
\description{
Create the WoEs in the data for a numeric variable
}
