% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sample.R
\name{sample_data}
\alias{sample_data}
\title{Sample a dataset}
\usage{
sample_data(ds, stratifiers = NULL, train_perc = 0.5)
}
\arguments{
\item{ds}{A dataset to be divided into train and test samples}

\item{stratifiers}{A character vector indicating which variables should be
to stratify the samples}

\item{train_perc}{Percentage of records which should be in the training
dataset}
}
\value{
A list of two components, the first component is the training
    sample and the other the test sample
}
\description{
Sample a dataset
}
