% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sample.R
\name{sample_server}
\alias{sample_server}
\title{Sample data module server-side processing}
\usage{
sample_server(input, output, session, modeldata, modeldataspecs)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{modeldata}{data frame containing the credit scoring data}

\item{modeldataspecs}{data specification table for credit scoring data}
}
\value{
list with following components
\describe{
  \item{goodbad_var}{good/bad variable}
  \item{stratification_vars}{stratification variables}
  \item{modeltraintest}{list of training and test data}
}
}
\description{
Sample data module server-side processing
}
