% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sample.R
\name{sample_ui}
\alias{sample_ui}
\title{Interface to sample the data based on data specifications}
\usage{
sample_ui(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Interface to sample the data based on data specifications
}
