% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_scorecardbuilder.R
\name{scorecardbuilder}
\alias{scorecardbuilder}
\title{Launches the scorecardbuilder application}
\usage{
scorecardbuilder()
}
\value{
Shiny application object
}
\description{
Launches the scorecardbuilder application
}
