% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validation.R
\name{validation_metrics_table}
\alias{validation_metrics_table}
\title{Create underlying table for validation metrics}
\usage{
validation_metrics_table(df, score = "score", pbad = "p_bad",
  gb = "f_good", n_groups = 20, brks = NULL)
}
\arguments{
\item{df}{Dataset containing the required variables}

\item{score}{Score variable}

\item{pbad}{Probability of bad}

\item{gb}{Actual good/bad flag with value 1 for goods}

\item{n_groups}{Number of groups to divide the data while calculating metrics}

\item{brks}{Pre-determined set of breaks to cut the score}
}
\value{
a list of the metrics table along with the breaks for the score
    bands
}
\description{
Create underlying table for validation metrics
}
