% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validation.R
\name{validation_server}
\alias{validation_server}
\title{Validation module server-side processing}
\usage{
validation_server(input, output, session, scorecard, goodbad_var)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{scorecard}{scorecard object created in model module}

\item{goodbad_var}{good/bad variable}
}
\value{
list with following components
\describe{
  \item{validation}{Object with the validation datasets}
}
}
\description{
Validation module server-side processing
}
