% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validation.R
\name{validation_ui}
\alias{validation_ui}
\title{Interface to display scorecard validation metrics}
\usage{
validation_ui(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Interface to display scorecard validation metrics
}
