% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_scorecardbuilder.R
\name{scorecardbuilder}
\alias{scorecardbuilder}
\title{launches the shinyAppDemo app}
\usage{
scorecardbuilder()
}
\value{
shiny application object
}
\description{
launches the shinyAppDemo app
}
